+++ toolkit/utilitynetworks/src/androidTest/java/com/arcgismaps/toolkit/utilitynetworks/ExampleInstrumentedTest.kt
@@ -16,7 +16,7 @@
  *
  */
 
-package com.arcgismaps.toolkit.utilitynetwork
+package com.arcgismaps.toolkit.utilitynetworks
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
